+++ cloudbuild.yaml
@@ -2,7 +2,7 @@ steps:
   - name: 'gcr.io/$PROJECT_ID/scala-sbt'
     args:
       - |
-        test; docker:publishLocal
+        test; resourceValidator/docker:publishLocal
   - name: 'gcr.io/cloud-builders/docker'
     args: [ 'image', 'tag', 'us.gcr.io/broad-dsp-gcr-public/resource-validator:latest', 'us.gcr.io/broad-dsp-gcr-public/resource-validator:$SHORT_SHA']
 images: [
