+++ resource-validator/src/main/scala/com/broadinstitute/dsp/resourceValidator/DeletedRuntimeChecker.scala
@@ -1,9 +1,12 @@
 package com.broadinstitute.dsp
-package zombieMonitor
+package resourceValidator
 
 import cats.effect.{Concurrent, Timer}
 import cats.implicits._
 import cats.mtl.ApplicativeAsk
+import com.broadinstitute.dsp
+import com.broadinstitute.dsp.CloudService.{Dataproc, Gce}
+import com.broadinstitute.dsp.{RuntimeChecker, RuntimeCheckerDeps}
 import io.chrisdavenport.log4cats.Logger
 import org.broadinstitute.dsde.workbench.google2.{DataprocClusterName, InstanceName}
 import org.broadinstitute.dsde.workbench.model.TraceId
@@ -21,17 +24,17 @@ object DeletedRuntimeChecker {
 
       override def dependencies: RuntimeCheckerDeps[F] = deps
 
-      def checkRuntimeStatus(runtime: Runtime,
-                             isDryRun: Boolean)(implicit ev: ApplicativeAsk[F, TraceId]): F[Option[Runtime]] =
+      def checkRuntimeStatus(runtime: dsp.Runtime,
+                             isDryRun: Boolean)(implicit ev: ApplicativeAsk[F, TraceId]): F[Option[dsp.Runtime]] =
         runtime.cloudService match {
-          case CloudService.Dataproc =>
+          case Dataproc =>
             checkClusterStatus(runtime, isDryRun)
-          case CloudService.Gce =>
+          case Gce =>
             checkGceRuntimeStatus(runtime, isDryRun)
         }
 
-      def checkClusterStatus(runtime: Runtime,
-                             isDryRun: Boolean)(implicit ev: ApplicativeAsk[F, TraceId]): F[Option[Runtime]] =
+      def checkClusterStatus(runtime: dsp.Runtime,
+                             isDryRun: Boolean)(implicit ev: ApplicativeAsk[F, TraceId]): F[Option[dsp.Runtime]] =
         for {
           clusterOpt <- dependencies.dataprocService
             .getCluster(runtime.googleProject, regionName, DataprocClusterName(runtime.runtimeName))
@@ -43,9 +46,9 @@ object DeletedRuntimeChecker {
                 .deleteCluster(runtime.googleProject, regionName, DataprocClusterName(runtime.runtimeName))
                 .void
           }
-        } yield clusterOpt.fold(none[Runtime])(_ => Some(runtime))
+        } yield clusterOpt.fold(none[dsp.Runtime])(_ => Some(runtime))
 
-      def checkGceRuntimeStatus(runtime: Runtime, isDryRun: Boolean): F[Option[Runtime]] =
+      def checkGceRuntimeStatus(runtime: dsp.Runtime, isDryRun: Boolean): F[Option[dsp.Runtime]] =
         for {
           runtimeOpt <- dependencies.computeService
             .getInstance(runtime.googleProject, zoneName, InstanceName(runtime.runtimeName))
@@ -58,8 +61,8 @@ object DeletedRuntimeChecker {
                   .deleteInstance(runtime.googleProject, zoneName, InstanceName(runtime.runtimeName))
                   .void
           }
-        } yield runtimeOpt.fold(none[Runtime])(_ => Some(runtime))
+        } yield runtimeOpt.fold(none[dsp.Runtime])(_ => Some(runtime))
 
-      override def runtimesToScan: fs2.Stream[F, Runtime] = dbReader.getDeletedRuntimes
+      override def runtimesToScan: fs2.Stream[F, dsp.Runtime] = dbReader.getDeletedRuntimes
     }
 }
