+++ resource-validator/src/test/scala/com/broadinstitute/dsp/zombieMonitor/mocks.scala
@@ -3,6 +3,7 @@ package zombieMonitor
 
 import fs2.Stream
 import cats.effect.IO
+import com.broadinstitute.dsp.resourceValidator.DbReader
 
 class FakeDbReader extends DbReader[IO] {
   override def getDeletedRuntimes: Stream[IO, Runtime] = Stream.empty
